Remove the e2eIntegrationTests file and split its functions out to other tests
e2eIntegrationTests was too large and outdated
To accommodate the Mocha update from 2.5.3 to 4.0.0

OHM-499
